build benchmarks in a different executable to unit tests
