+++ test/benchmarks/threads/test_multicore.h
@@ -1,6 +1,8 @@
 #pragma once
-#include <iostream>
-
+#include <algorithm>
+#include <cmath>
+#include <string>
+#include <vector>
 
 namespace nrn::test {
 extern int MAX_PROCESSORS;
